feat: overlapping labels exercise
Simulate issue when training machine learning models using targets that
are returns that overlap in time. The core issue is that these labels are
correlated and in fact redundant. We'll see what impact this has on our machine
learning algorithm, and we'll have the opportunity to implement some of the
solutions to the problem that we described in the lectures, and will later
encounter in the project.