feat: project 08 - backtesting
Add Jupyter notebook to build a fairly realistic backtester that uses the Barra data.
The backtester performs portfolio optimization that includes transaction costs,
and it is implemented with computational efficiency in mind, to allow for a
reasonably fast backtest. Performance attribution is used to identify the
major drivers of the portfolio's profit-and-loss (PnL).